добавлена работа с фотографиями

+++ src/main/java/com/example/sparkle/sparkle/controller/UserController.java
@@ -1,13 +1,16 @@
 package com.example.sparkle.sparkle.controller;
 
-import com.example.sparkle.sparkle.dto.user.*;
+import com.example.sparkle.sparkle.builder.BuilderUser;
+import com.example.sparkle.sparkle.dto.user.UserDtoRegister;
+import com.example.sparkle.sparkle.dto.user.UserDtoUpdate;
 import com.example.sparkle.sparkle.model.User;
-import com.example.sparkle.sparkle.repository.UserRepository;
 import com.example.sparkle.sparkle.service.UserService;
+import com.example.sparkle.sparkle.validator.ValidatorUser;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Valid;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
+import jakarta.validation.constraints.Min;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -24,13 +27,11 @@ import java.util.*;
 public class UserController {
 
     private final UserService userService;
-    private final UserRepository userRepository;
+
 
     @Autowired
-    public UserController(UserService userService, UserRepository userRepository) {
+    public UserController(UserService userService) {
         this.userService = userService;
-        this.userRepository = userRepository;
-
     }
 
     /**
@@ -39,13 +40,14 @@ public class UserController {
      */
     @PostMapping("/register/next-page")
     public ResponseEntity<?> registerUser(@Valid @RequestBody UserDtoRegister userDtoRegister) {
-        User user = userService.getUserByEmail(userDtoRegister.getEmail());
+        Optional<User> user = userService.getUserByEmail(userDtoRegister.getEmail());
         try {
-            if (user != null && user.getEmail().equals(userDtoRegister.getEmail())) {
-                log.debug("Попытка зарегистрироваться по существующему email = {}", userDtoRegister.getEmail());
+            if (user.isPresent() && user.get().getEmail().equals(userDtoRegister.getEmail())) {
+                log.info("Попытка зарегистрироваться по существующему email = {}", userDtoRegister.getEmail());
                 return ResponseEntity.badRequest().body("Пользователь с таким email уже зарегистрирован");
             }
-            return ResponseEntity.ok(userService.registerUser(userDtoRegister));
+            user = Optional.ofNullable(userService.registerUser(BuilderUser.userBuilder(userDtoRegister)));
+            return ResponseEntity.ok(BuilderUser.userDtoRegisterBuilder(user.orElseThrow()));
         } catch (MethodArgumentNotValidException e) {
             Map<String, String> errors = new HashMap<>();
             for (FieldError error : e.getBindingResult().getFieldErrors()) {
@@ -55,67 +57,46 @@ public class UserController {
         }
     }
 
-    /**
-     * Аутентификация пользователя
-     * Если пользователь уже зарегистрировался и хочет войти под своим email и паролем
-     */
-    @GetMapping("/authenticate")
-    public ResponseEntity<?> authenticateUser(@Valid @RequestBody UserDtoAuthenticate userDtoAuthenticate) {
-        User user = userService.getUserByEmail(userDtoAuthenticate.getEmail());
-        if (user == null) {
-            log.debug("Пользователь с email = {} не найден", userDtoAuthenticate.getEmail());
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Пользователь не найден");
-        }
-        if (!user.getPassword().equals(userDtoAuthenticate.getPassword())) {
-            log.debug("Пароль введен не верно");
-            return new ResponseEntity<>("Пароль введен не верно", HttpStatus.BAD_REQUEST);
-        }
-        log.info("Успешный вход пользователя");
-        return ResponseEntity.ok(userService.authenticateUser(user));
-    }
-
 
     /**
      * Получение информации о текущем авторизованном пользователе
      */
     @GetMapping("/profile/{userId}")
-    public ResponseEntity<?> getCurrentUserProfile(@PathVariable Long userId) {
-        UserDtoGet userDtoGet = userService.getCurrentUserProfile(userId);
-        if (userDtoGet != null) {
-            log.info("Пользователь с ID = {} получен", userId);
-            return ResponseEntity.ok(userDtoGet);
+    public ResponseEntity<?> getCurrentUserProfile(@PathVariable @Min(1) Long userId) {
+        Optional<User> user = userService.getCurrentUserProfile(userId);
+        ResponseEntity<?> validationResult = ValidatorUser.userNotFoundAndForbidden(
+                user.orElse(null),
+                userId);
 
+        if (!validationResult.getStatusCode().is2xxSuccessful()) {
+            return validationResult;
         }
-        log.debug("Пользователь с ID = {} не найден", userId);
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Пользователь не найден");
+        log.info("Пользователь с ID = {} получен", userId);
+        return ResponseEntity.ok(user);
     }
 
     /**
      * Редактирование профиля пользователя
      */
     @PatchMapping("/update-profile/{userId}")
-    public ResponseEntity<?> updateUserProfile(@PathVariable Long userId,
+    public ResponseEntity<?> updateUserProfile(@PathVariable @Min(1) Long userId,
                                                @Valid @RequestBody UserDtoUpdate userDtoUpdate) {
         try {
-            Optional<User> optionalUser = userRepository.findById(userId);
-            if (!optionalUser.isPresent()) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Пользователь не найден");
+            Optional<User> user = userService.getUserById(userId);
+            ResponseEntity<?> validationResult = ValidatorUser.userNotFoundAndForbidden(
+                    user.orElse(null),
+                    userId);
+
+            if (!validationResult.getStatusCode().is2xxSuccessful()) {
+                return validationResult;
             }
-            userDtoUpdate.setId(userId);
-            UserDtoRegister validationDto = new UserDtoRegister();
-            validationDto.setUsername(optionalUser.get().getUsername());
-            validationDto.setEmail(optionalUser.get().getEmail());
-            validationDto.setPassword(optionalUser.get().getPassword());
-            validationDto.setBirthDate(optionalUser.get().getBirthDate());
-            validationDto.setGender(optionalUser.get().getGender());
+
+            UserDtoRegister validationDto = BuilderUser.userDtoRegisterBuilder(user.orElseThrow());
             if (userDtoUpdate.getUsername() != null) {
                 validationDto.setUsername(userDtoUpdate.getUsername());
             }
             if (userDtoUpdate.getEmail() != null) {
                 validationDto.setEmail(userDtoUpdate.getEmail());
-            }
-            if (userDtoUpdate.getPassword() != null) {
-                validationDto.setPassword(userDtoUpdate.getPassword());
             }
             if (userDtoUpdate.getBirthDate() != null) {
                 validationDto.setBirthDate(userDtoUpdate.getBirthDate());
@@ -123,8 +104,7 @@ public class UserController {
             if (userDtoUpdate.getGender() != null) {
                 validationDto.setGender(userDtoUpdate.getGender());
             }
-            log.debug("Валидация данных");
-            log.debug(">>>>>> " + validationDto);
+            log.info("Валидация данных");
             Validator validator = Validation.buildDefaultValidatorFactory().getValidator();
 
             if (userDtoUpdate.getUsername() == null || userDtoUpdate.getUsername().isBlank()) {
@@ -133,8 +113,7 @@ public class UserController {
                 );
                 if (!violations.isEmpty()) {
                     return ResponseEntity.badRequest()
-                            .body(violations.iterator().next().getMessage())
-                            ;
+                            .body(violations.iterator().next().getMessage());
                 }
             }
 
@@ -149,24 +128,13 @@ public class UserController {
                 }
             }
 
-            if (userDtoUpdate.getPassword() != null || userDtoUpdate.getPassword() == null) {
-                Set<ConstraintViolation<UserDtoRegister>> violations = validator.validateProperty(
-                        validationDto, "password"
-                );
-                if (!violations.isEmpty()) {
-                    return ResponseEntity.badRequest()
-                            .body(violations.iterator().next().getMessage())
-                            ;
-                }
-            }
             if (userDtoUpdate.getGender() != null || userDtoUpdate.getGender() == null) {
                 Set<ConstraintViolation<UserDtoRegister>> violations = validator.validateProperty(
                         validationDto, "gender"
                 );
                 if (!violations.isEmpty()) {
                     return ResponseEntity.badRequest()
-                            .body(violations.iterator().next().getMessage())
-                            ;
+                            .body(violations.iterator().next().getMessage());
                 }
             }
             if (userDtoUpdate.getBirthDate() != null || userDtoUpdate.getBirthDate() == null) {
@@ -175,25 +143,24 @@ public class UserController {
                 );
                 if (!violations.isEmpty()) {
                     return ResponseEntity.badRequest()
-                            .body(violations.iterator().next().getMessage())
-                            ;
+                            .body(violations.iterator().next().getMessage());
                 }
             }
 
-            int affectedRows = userService.updateUserProfile(userDtoUpdate);
+            int affectedRows = userService.updateUserProfile(BuilderUser.userDtoUpdateBuilder(userDtoUpdate));
 
             if (affectedRows == 0) {
                 log.warn("Обновление пользователя с ID = {} не выполнено - запись не найдена", userId);
                 return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Пользователь не найден");
             }
-            log.debug("Обновление пользователя с ID = {} прошло успешно", userDtoUpdate.getId());
-
-            return ResponseEntity.ok(userService.getUserById(userId));
+            log.info("Обновление пользователя с ID = {} прошло успешно", userDtoUpdate.getId());
+            user = userService.getUserById(userId);
+            return ResponseEntity.ok(BuilderUser.userDtoRegisterBuilder(user.orElseThrow()));
 
 
         } catch (RuntimeException e) {
             if (e.getMessage().contains("unique")) {
-                log.debug("Пользователь с email = {} уже зарегистрирован {} ", userDtoUpdate.getEmail(), e.getMessage());
+                log.info("Пользователь с email = {} уже зарегистрирован {} ", userDtoUpdate.getEmail(), e.getMessage());
                 return new ResponseEntity<>("Пользователь с таким email уже зарегистрирован", HttpStatus.BAD_REQUEST);
             } else {
                 log.debug(e.getMessage());
@@ -209,36 +176,53 @@ public class UserController {
      */
     @GetMapping
     public ResponseEntity<?> getUserAll() {
-        List<UserDtoRegisterGet> users = userService.getUserAll();
+        List<User> users = userService.getUserAll();
         if (users.isEmpty()) {
             log.info("Пользователи отсутствуют");
             return ResponseEntity.noContent().build();
 
         }
-        return ResponseEntity.ok(users);
+        return ResponseEntity.ok(users.stream().map(BuilderUser::userDtoRegisterBuilder));
     }
+
     /**
      * Получение пользователя по ID
      */
     @GetMapping("/{userId}")
-    public ResponseEntity<?> getUserById(@PathVariable Long userId) {
-        UserDtoRegisterGet userDtoRegisterGet = userService.getUserById(userId);
-        if (userDtoRegisterGet == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Пользователь не найден");
+    public ResponseEntity<?> getUserById(@PathVariable @Min(1) Long userId) {
+        Optional<User> user = userService.getUserById(userId);
+        ResponseEntity<?> validationResult = ValidatorUser.userNotFoundAndForbidden(
+                user.orElse(null),
+                userId);
+
+        if (!validationResult.getStatusCode().is2xxSuccessful()) {
+            return validationResult;
         }
-        return ResponseEntity.ok(userDtoRegisterGet);
+        return ResponseEntity.ok(BuilderUser.userDtoRegisterBuilder(user.orElseThrow()));
     }
 
     /**
      * Удаление пользователя по ID
      */
     @DeleteMapping("/{userId}")
-    public void deleteUserById(@PathVariable Long userId) {
-        UserDtoRegisterGet userDtoRegisterGet = userService.getUserById(userId);
-        if (userDtoRegisterGet == null) {
-            ResponseEntity.status(HttpStatus.NOT_FOUND).body("Пользователь не найден");
+    public ResponseEntity<?> deleteUserById(@PathVariable @Min(1) Long userId) {
+        Optional<User> user = userService.getUserById(userId);
+        ResponseEntity<?> validationResult = ValidatorUser.userNotFoundAndForbidden(
+                user.orElse(null),
+                userId);
+
+        if (!validationResult.getStatusCode().is2xxSuccessful()) {
+            return validationResult;
         }
+
         userService.deleteUserById(userId);
+        user = userService.getUserById(userId);
+        if (user.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        log.debug("Ошибка удаления пользователя");
+        return new ResponseEntity<>(HttpStatus.CONFLICT);
+
     }
 
 }
